Add support for Docker API version 1.20
Previously this app was built to use Docker API version 1.8.  This
was fine and dandy up until docker 1.7 when the minimum version became
restricted to 1.12.  I chose 1.20 since this is the API released with
docker 1.8, which seems to be widely available and supported on most
modern enterprise systems (like CentOS).

Also updated all tests to ensure they worked.  And my editor was
nice enough to strip out all trailing whitespace.  Sorry for the
mess there.